+++ anymail/inbound.py
@@ -1,6 +1,7 @@
 import re
 import warnings
 from base64 import b64decode
+from email.feedparser import BytesFeedParser
 from email.message import EmailMessage
 from email.parser import BytesParser, Parser
 from email.policy import default as default_policy
@@ -257,6 +258,15 @@ class AnymailInboundMessage(EmailMessage):
             return BytesParser(cls, policy=default_policy).parse(fp)
         else:
             return Parser(cls, policy=default_policy).parse(fp)
+
+    @classmethod
+    def parse_raw_mime_chunks(cls, chunks):
+        """Returns a new AnymailInboundMessage parsed from (streamed) chunks"""
+        parser = BytesFeedParser(cls, policy=default_policy)
+        for chunk in chunks:
+            if chunk:
+                parser.feed(chunk)
+        return parser.close()
 
     @classmethod
     def construct(