+++ CHANGELOG.rst
@@ -35,6 +35,9 @@ Features
 
 * **Mailtrap:** Support signature verification for tracking event webhooks.
 
+* **Mailtrap:** Add support for inbound email. See
+  `docs <https://anymail.dev/en/latest/esps/mailtrap/#inbound-webhook>`.
+
 * **Postmark:** Add support for Postmark's bulk API. See
   `Using Postmark's bulk API
   <https://anymail.dev/en/latest/esps/postmark/#using-postmark-s-bulk-api>`_.